Fix linter issues by splitting up Signal helpers & more
