+++ .swiftlint.yml
@@ -5,6 +5,7 @@ disabled_rules: # rule identifiers turned on by default to exclude from running
   - trailing_comma
   - function_parameter_count
   - closure_parameter_position
+  - nesting
 opt_in_rules: # some rules are turned off by default, so you need to opt-in
   - empty_count # Find all the available rules by running: `swiftlint rules`
